requirements.txt
aioredis==2.0.1
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.3.0
black==24.1.1
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
click==8.2.0
cryptography==45.0.1
distlib==0.3.9
ecdsa==0.19.1
fastapi==0.115.12
filelock==3.18.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.10
idna==3.10
iniconfig==2.1.0
iso8601==2.1.0
isort==5.13.2
loguru==0.7.3
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pre-commit==3.6.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyJWT==2.10.1
pypika-tortoise==0.5.0
pytest==7.4.2
pytest-asyncio==0.23.4
python-dotenv==1.1.0
python-jose==3.3.0
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tortoise-orm==0.25.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==1.0.5
websockets==15.0.1
